UI Change: Rebind of ‘%’ in Casual Calc Main Menu.
In prior releases of Casual ≤ v2.2.3, the Calc main menu (`casual-calc-tmenu`)
bound the key ‘%’ to the command `calc-percent`. This however breaks with the
more common calculator convention of mapping the ‘%’ behavior to actually
calculating a percentage given two numbers, the value and the percent rate. Also
it is my opinion that the behavior of `calc-percent` is not compelling enough to
merit placement on the Calc main menu.

This commit changes this behavior so that pressing the ‘%’ key will now
calculate a percentage. For example, if `85` is stored in (2:) and `40` in (1:),
pressing ‘%’ will leave the result `34` in (1:).

The Financial main menu (`casual-calc-financial-tmenu`) has also been modified
accordingly. Access to the command `calc-percent` has been moved to this menu
with the binding ‘P’.

Note that only in both `casual-calc-tmenu` and `casual-calc-financial-tmenu`,
the ‘%’ key is reserved for percentage calculation. In all other Casual Calc
menus and Calc itself, the ‘%’ key is reserved to be the modulo operator.