Fix I-Search Toggle Behavior in `casual-isearch-tmenu`
Prior to this change, search toggle menu items (regexp, symbol, word)
erroneously invoked `isearch-edit-string`, forcing a recursive edit. This
created a number of errors, namely:

- During recursive edit, the binding to `casual-isearch-tmenu` was not
available, blocking a path to change the search mode.
- Recursive edit itself is surprising to users not familiar with it.

This fix changes to avoid invoking `isearch-edit-string` in the first place.

New behavior is to invoke the I-Search toggle command directly. In addition, the
`:transient` key is set to `nil`, to return focus back to the I-Search
minibuffer prompt. This allows the binding to `casual-isearch-tmenu` to be
retained.

Descriptions for toggle switches have been changed to use checkboxes for
improved usability.

Update documentation, including revert of auto-versioning document. This needs
to wait for command line generation of the texinfo file.

Fix Replace section for read-only behavior.